+++ internal/pet/config.go
@@ -76,7 +76,7 @@ type Config struct {
 	IconTheme    IconTheme         `json:"icon_theme,omitempty"`
 	Separator    string            `json:"separator"`
 	Lines        []string          `json:"lines,omitempty"`
-	LineColors   [][]uint8         `json:"line_colors,omitempty"`
+	LineColors   [][]Color         `json:"line_colors,omitempty"`
 	DisplayMode  DisplayMode       `json:"display_mode,omitempty"`
 	WrapCommand  string            `json:"wrap_command,omitempty"`
 	BarStyle     BarStyle          `json:"bar_style,omitempty"`